WIP element selection

+++ app/src/main/java/cz/cvut/arfittingroom/draw/model/element/impl/Curve.kt
@@ -14,7 +14,7 @@ class Curve(
         canvas.drawPath(path, paint)
     }
 
-    override fun doIntersect(x: Int, y: Int): Boolean {
+    override fun doIntersect(x: Float, y: Float): Boolean {
         TODO("Not yet implemented")
     }
 }
