+++ app/src/main/java/cz/cvut/arfittingroom/draw/model/element/impl/Image.kt
@@ -26,7 +26,7 @@ class Image(
         canvas.drawBitmap(imageBitmap, matrix, null)
     }
 
-    override fun doIntersect(x: Int, y: Int): Boolean {
+    override fun doIntersect(x: Float, y: Float): Boolean {
         TODO("Not yet implemented")
     }
 }
